requirements.txt
# ===== CORE FLASK FRAMEWORK =====
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Werkzeug==3.1.3

# ===== DATABASE AND ORM =====
SQLAlchemy==2.0.37
alembic==1.14.1
greenlet==3.1.1

# ===== FORM HANDLING AND VALIDATION =====
WTForms==3.2.1
email_validator==2.2.0

# ===== TEMPLATE ENGINE =====
Jinja2==3.1.5
MarkupSafe==3.0.2

# ===== IMAGE PROCESSING AND OCR =====
Pillow==11.1.0
pytesseract==0.3.13

# ===== DOCUMENT PROCESSING =====
python-docx==1.2.0
docx-mailmerge2==0.9.0
lxml==6.0.0

# ===== QR CODE AND PDF GENERATION =====
qrcode[pil]==8.0
reportlab==4.2.5
pdf2image==1.17.0

# ===== SECURITY AND UTILITIES =====
python-dotenv==1.0.1
itsdangerous==2.2.0
blinker==1.9.0

# ===== CORE PYTHON UTILITIES =====
click==8.1.8
dnspython==2.7.0
idna==3.10
Mako==1.3.8
packaging==24.2
setuptools>=65.0.0
typing_extensions==4.12.2

# ===== DEVELOPMENT AND TESTING (Optional) =====
# Uncomment for development environment
# pytest==8.0.0
# pytest-flask==1.3.0
# black==24.0.0
# flake8==7.0.0
# isort==5.13.0

# ===== PRODUCTION DEPLOYMENT (Optional) =====
# Uncomment for production deployment
# gunicorn==21.2.0
# psycopg2-binary==2.9.9  # For PostgreSQL
# redis==5.0.0  # For caching/sessions
# celery==5.3.0  # For background tasks

# ===== MONITORING AND LOGGING (Optional) =====
# Uncomment for production monitoring
# sentry-sdk[flask]==1.40.0
# python-json-logger==2.0.7
scikit-learn
